requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
colorama==0.4.1
distlib==0.3.0
Django==3.0.7
django-filter==2.3.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pipenv==2018.11.26
pycodestyle==2.5.0
pylint==2.4.4
pytz==2020.1
six==1.13.0
sqlparse==0.3.1
typed-ast==1.4.0
virtualenv==20.0.16
virtualenv-clone==0.5.4
whitenoise==5.1.0
wrapt==1.11.2
zipp==3.1.0
